requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
chromadb==1.0.15
google-adk==1.3.0
litellm==1.75.0